[TRTLLM-13212][refactor] Centralize sampling logic, split backends into isolated modules
- Merge sampling_utils_flashinfer.py into sampling_utils.py (deleted former)
- Move _StrategyImpls and FlashInferGroupedStrategySampler to sampling_utils.py
- Wrap all flashinfer kernel calls in backend_flashinfer.py wrapper functions
- Move _st_* helpers to backend_torch.py (drop prefix, rename to forward_native_sampling etc.)
- Move rejection_sampling_one_model to speculative/interface.py (draft acceptance logic)
- Remove use_flashinfer param from sampling_batch_spec_dec_one_model (auto-detect)
- Add sanitize_top_k (top_k<=0 means keep-all) and apply it in the spec-dec sampling facade
- Add torch fallback path to sampling_batch_spec_dec_one_model_for_rejection
- Fix top_p renormalization: drop in-place cumsum(out=probs_sorted) that corrupted the denominator
- Fix sampler.py import: FlashInferGroupedStrategySampler now from sampling_utils
- Move StrategyMetadata to backend_torch.py so BeamSearchMetadata can inherit it
- Use explicit re-export form for _Fusions, get_rejected_indices, sample_rejected
- Add type annotations to _StrategyImpls subclass sample() overrides and from_strategies
- Update sampler unit tests for the new _bound_backend dispatch + module locations
