+++ tensorrt_llm/_torch/pyexecutor/sampler.py
@@ -92,7 +92,6 @@ from .sampling_utils import (
     GREEDY,
     BeamSearchMetadata,
     GenericStrategyKeyType,
-    SimpleGroupedStrategySampler,
     Strategy,
     StrategyMetadata,
     UtilsSamplingParams,
@@ -113,8 +112,6 @@ if TYPE_CHECKING:
     from transformers import PretrainedConfig
 
     from tensorrt_llm._torch.models.modeling_utils import DecoderModel, DecoderModelForCausalLM
-
-    from .sampling_utils_flashinfer import FlashInferGroupedStrategySampler
 
     _ModelType = TypeVar("_ModelType", bound=DecoderModel)
     _ConfigType = TypeVar("_ConfigType", bound=PretrainedConfig)
@@ -2350,15 +2347,14 @@ class TorchSampler(Sampler[SampleStateTorch], AsyncWorkerMixin):
         self.LOGPROBS_SHAPE = (self.max_num_sequences, self.max_beam_width, self.max_tokens)
         self.TOPK_LOGPROBS_SHAPE = (self.max_num_sequences, self.max_tokens, self.max_topk_logprobs)
 
-        self._grouped_sampler_cls: (
-            Type["FlashInferGroupedStrategySampler"] | Type[SimpleGroupedStrategySampler]
-        )
-        if IS_FLASHINFER_AVAILABLE and not args.disable_flashinfer_sampling:
-            from .sampling_utils_flashinfer import FlashInferGroupedStrategySampler
+        from .sampling_utils import BoundSamplingBackend, SamplerConfig, resolve_sampling_backend
 
-            self._grouped_sampler_cls = FlashInferGroupedStrategySampler
-        else:
-            self._grouped_sampler_cls = SimpleGroupedStrategySampler
+        self._bound_backend: BoundSamplingBackend = resolve_sampling_backend(
+            torch.device("cuda"),
+            SamplerConfig(
+                use_flashinfer=IS_FLASHINFER_AVAILABLE and not args.disable_flashinfer_sampling,
+            ),
+        )
 
         # AutoDeploy build creates the sampler in inference mode,
         # which would disallow in-place mutating of new_tokens.
@@ -4050,7 +4046,7 @@ class TorchSampler(Sampler[SampleStateTorch], AsyncWorkerMixin):
         grouped_requests = self._request_grouper.group_requests_by_strategy_key(
             requests,
             pin_memory=prefer_pinned(),
-            strategy_to_key=self._grouped_sampler_cls.strategy_grouping_key,
+            strategy_to_key=self._bound_backend.strategy_grouping_key,
             seq_slots=seq_slots,
             vocab_size=logits_cuda.size(1),  # Dummy value; strategy should already be cached
         )
@@ -4059,7 +4055,7 @@ class TorchSampler(Sampler[SampleStateTorch], AsyncWorkerMixin):
             grouped_requests,
             seq_slots,
             seq_lens,
-            get_metadata_type_for_group_fn=self._grouped_sampler_cls.get_metadata_type_for_group,
+            get_metadata_type_for_group_fn=self._bound_backend.get_metadata_type_for_group,
             seq_slots_cuda=seq_slots_cuda,
             seq_lens_cuda=seq_lens_cuda,
         )
@@ -4185,7 +4181,7 @@ class TorchSampler(Sampler[SampleStateTorch], AsyncWorkerMixin):
             ]
 
             group_next_tokens_cuda, group_softmax_cuda, group_temperature_cuda = (
-                self._grouped_sampler_cls.sample_grouped_strategies(
+                self._bound_backend.sample_grouped_strategies(
                     strategy_key,
                     group_strategies_per_step,
                     group_logits_cuda,
